+++ lib/router/src/ui/page_not_found.dart
@@ -7,7 +7,7 @@ class PageNotFound extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Page Not Found'.tr())),
+      appBar: AppBar(title: Text('nylo.page_not_found.title'.tr())),
       body: Center(
         child: Container(
           margin: const EdgeInsets.all(8.0),
@@ -17,13 +17,13 @@ class PageNotFound extends StatelessWidget {
               const Icon(Icons.info_outline, size: 80.0, color: Colors.black38),
               const SizedBox(height: 24.0),
               Text(
-                'Sorry, the page you have requested is not available.'.tr(),
+                'nylo.page_not_found.message'.tr(),
                 textAlign: TextAlign.center,
                 style: const TextStyle(fontSize: 16.0),
               ),
               ElevatedButton(
                 onPressed: () => Navigator.pop(context),
-                child: Text("Go back".tr()),
+                child: Text("nylo.page_not_found.go_back".tr()),
               ),
             ],
           ),